+++ tests/conftest.py
@@ -263,3 +263,15 @@ def reset_module_state():
     db_module.close_db_connection()
     # Clear migrated databases set after test
     db_module._reset_migration_state_for_tests()
+
+
+@pytest.fixture
+def mock_sqlite_select_db(monkeypatch):
+    """Fixture to mock sqlite.select_db."""
+    from unittest.mock import MagicMock
+
+    from spotfm import sqlite
+
+    mock_db = MagicMock()
+    monkeypatch.setattr(sqlite, "select_db", mock_db)
+    return mock_db
